Login 페이지 리팩토링 : ScriptBar 분리
- scriptBar 컴포넌트를 분리
