+++ frontend/src/components/login/index.ts
@@ -1,2 +1,3 @@
 export { default as TypingAnimeComponent } from "./TypingAnimeComponent";
 export { default as GithubLoginButton } from "./GithubLoginButton";
+export { default as ScriptBar } from "./ScriptBar";
